project3/backend/00-data-ingestion/polygon-historical-downloader/requirements.txt
# Polygon.io Official Client
polygon-api-client==1.14.1

# Message Queues
nats-py==2.7.0
aiokafka==0.10.0

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Async Support
aiohttp==3.9.3
asyncio==3.4.3

# Logging
python-json-logger==2.0.7

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Date/Time handling
python-dateutil==2.8.2

# Progress tracking
tqdm==4.66.1

# HTTP
requests==2.31.0

# Central Hub SDK (installed from local path in Dockerfile)
# httpx will be installed as SDK dependency
